+++ example/lib/main.dart
@@ -1,9 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:cellphone_validator/cellphone_validator.dart';
 
-
 void main() {
-
   runApp(const MyApp());
 }
 
@@ -13,14 +11,13 @@ class MyApp extends StatelessWidget {
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-    return
-      MaterialApp(
+    return MaterialApp(
       title: 'Flutter Demo',
       debugShowCheckedModeBanner: false,
       theme: ThemeData(
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
       ),
-      home:MyHomePage(title: 'Flutter Demo Home Page')
+      home: MyHomePage(title: 'Flutter Demo Home Page'),
     );
   }
 }
@@ -35,15 +32,30 @@ class MyHomePage extends StatefulWidget {
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  ValueNotifier<PhoneValidator> phoneValidator = ValueNotifier(PhoneValidator(lang: 'en'));
+  ValueNotifier<PhoneValidator> phoneValidator = ValueNotifier(
+    PhoneValidator(lang: 'en'),
+  );
   String selectedLanguage = 'en'; // Default language
-  final List<String> languages = ['ar','hi','id','it','ja','pt','en', 'es', 'fr','ko','de','ru', 'ur']; // Add more languages as needed
+  final List<String> languages = [
+    'ar',
+    'hi',
+    'id',
+    'it',
+    'ja',
+    'pt',
+    'en',
+    'es',
+    'fr',
+    'ko',
+    'de',
+    'ru',
+    'ur',
+  ]; // Add more languages as needed
   List<Country> countries = [];
 
   PhoneValidator phoneValidatorEn = PhoneValidator(lang: 'en');
   PhoneValidator phoneValidatorEs = PhoneValidator(lang: 'es');
   PhoneValidator phoneValidatorUr = PhoneValidator(lang: 'ur');
-
 
   @override
   void initState() {
@@ -51,54 +63,61 @@ class _MyHomePageState extends State<MyHomePage> {
     super.initState();
   }
 
-
-
   @override
   Widget build(BuildContext context) {
-    return ListenableBuilder(listenable: phoneValidator, builder: (context,_)=>
-        Scaffold(
-      appBar: AppBar(
-        centerTitle: true,
-        title: Text(widget.title),
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-        actions: [
-          languageDropdown()
-        ],
-      ),
-      body: Center(
-        child: Column(
-          children: [
-            FractionallySizedBox(
+    return ListenableBuilder(
+      listenable: phoneValidator,
+      builder: (context, _) => Scaffold(
+        appBar: AppBar(
+          centerTitle: true,
+          title: Text(widget.title),
+          backgroundColor: Theme.of(context).colorScheme.inversePrimary,
+          actions: [languageDropdown()],
+        ),
+        body: Center(
+          child: Column(
+            children: [
+              FractionallySizedBox(
                 widthFactor: 1,
-                child: PhoneInputSelectorView(phoneValidator: phoneValidator.value!)),
-            ListenableBuilder(
+                child: PhoneInputSelectorView(
+                  phoneValidator: phoneValidator.value!,
+                ),
+              ),
+              ListenableBuilder(
                 listenable: phoneValidator!.value.isValidPhoneNotifier,
                 builder: (context, _) {
                   return phoneValidator!.value.isValidPhoneNotifier.value
                       ? PhoneSummaryView(
                           phoneValidator: phoneValidator.value!,
-                          fullPhoneNumber: phoneValidator.value!.phone.replaceAll('+', ''),)
+                          fullPhoneNumber: phoneValidator.value!.phone
+                              .replaceAll('+', ''),
+                        )
                       : Text('Invalid phone');
-                }),
-            PhoneSummaryView(
-              phoneValidator: phoneValidatorEs,
-              fullPhoneNumber: '34612345678',),
-            PhoneAutoDetectView(phoneValidator: phoneValidatorEn, fullPhoneNumber: '',),
-            const Divider(),
-            const Text('PhoneCountryInput (Pre-fixed to Argentina):'),
-            PhoneCountryInput(
-              phoneValidator: phoneValidatorEs,
-              countryIsoCode: 'AR',
-            ),
-          ],
-
+                },
+              ),
+              PhoneSummaryView(
+                phoneValidator: phoneValidatorEs,
+                fullPhoneNumber: '34612345678',
+              ),
+              PhoneAutoDetectView(
+                phoneValidator: phoneValidatorEn,
+                fullPhoneNumber: '',
+              ),
+              const Divider(),
+              const Text('PhoneCountryInput (Pre-fixed to Argentina):'),
+              PhoneCountryInput(
+                phoneValidator: phoneValidatorEs,
+                countryIsoCode: 'AR',
+              ),
+            ],
+          ),
         ),
       ),
-    ));
+    );
   }
 
-  DropdownButton <String> languageDropdown(){
-   return DropdownButton<String>(
+  DropdownButton<String> languageDropdown() {
+    return DropdownButton<String>(
       value: phoneValidator.value.lang,
       icon: const Icon(Icons.language),
       onChanged: (String? newValue) {
@@ -112,9 +131,4 @@ class _MyHomePageState extends State<MyHomePage> {
       }).toList(),
     );
   }
-
 }
-
-
-
-